+++ mzcore/src/glycan/positioned_structure.rs
@@ -24,9 +24,9 @@ pub struct PositionedGlycanStructure {
     /// All branches
     pub branches: Vec<PositionedGlycanStructure>,
     /// The inner depth, the number of steps needed to get to the root / attached amino acid
-    pub inner_depth: u32,
+    pub inner_depth: u16,
     /// The outer depth, the number of steps needed to get to the closest leaf node
-    pub outer_depth: u32,
+    pub outer_depth: u16,
     /// The branches taken to get to this location (from the root) as the index in the branches and the index in the branches when sorted by mass.
     /// For a general glycan with a fucose on the first hexnac and a bisection after the core double
     /// hexnac + hex, this variable will contain an empty list for the root hexnac. For the fucose
